fixed some errors in sales_engine regarding repo creation, renamed 'customers' to 'customer'
