+++ lib/customer.rb
