+++ lib/invoice_item.rb
@@ -6,15 +6,15 @@ class InvoiceItem
               :id,
               :invoice_id,
               :created_at,
-              :item_id
+              :item_id,
               :updated_at
 
   def initialize(data, repository)
     @repository  = repository
-    @id          = data[:id]
-    @item_id     = data[:item_id]
-    @invoice_id  = data[:invoice_id]
-    @quantity    = data[:quantity]
+    @id          = data[:id].to_i
+    @item_id     = data[:item_id].to_i
+    @invoice_id  = data[:invoice_id].to_i
+    @quantity    = data[:quantity].to_i
     @unit_price  = data[:unit_price]
     @created_at  = data[:created_at]
     @updated_at  = data[:updated_at]
