+++ lib/merchant_repository.rb
@@ -19,4 +19,12 @@ class MerchantRepository
     "#<{self.class} #{merchants.size} rows>"
   end
 
+  def all
+    merchants
+  end
+
+  def random
+    merchants.sample
+  end
+
 end
